+++ lib/saml/complex_types/statement_abstract_type.rb
@@ -3,7 +3,7 @@ module Saml
     class StatementAbstractType
       include HappyMapper
 
-      register_namespace 'xsi', 'http://www.w3.org/2001/XMLSchema-instance'
+      register_namespace 'xsi', Saml::XSI_NAMESPACE
 
       tag 'Statement'
       namespace 'saml'