+++ lib/saml/xml_helpers.rb
@@ -27,8 +27,8 @@ module Saml
 
       builder = Nokogiri::XML::Builder.new(:encoding => "UTF-8")
       builder.Envelope(:'xmlns:soapenv' => "http://schemas.xmlsoap.org/soap/envelope/",
-                       :'xmlns:xsd'     => "http://www.w3.org/2001/XMLSchema",
-                       :'xmlns:xsi'     => "http://www.w3.org/2001/XMLSchema-instance") do |xml|
+                       :'xmlns:xsd'     => Saml::XS_NAMESPACE,
+                       :'xmlns:xsi'     => Saml::XSI_NAMESPACE) do |xml|
         builder.parent.namespace = builder.parent.namespace_definitions.find { |n| n.prefix == 'soapenv' }
         builder.Body do
           builder.parent.add_child body.doc.root