+++ spec/factories/all.rb
@@ -27,7 +27,7 @@ class StatementDummy
   include HappyMapper
 
   register_namespace 'xacml-saml', "urn:oasis:xacml:2.0:saml:assertion:schema:os"
-  register_namespace 'xsi', 'http://www.w3.org/2001/XMLSchema-instance'
+  register_namespace 'xsi', Saml::XSI_NAMESPACE
 
   tag 'Statement'
   namespace 'saml'
@@ -130,6 +130,9 @@ FactoryGirl.define do
 
   factory :attribute, :class => Saml::Elements::Attribute do
 
+  end
+
+  factory :attribute_value, :class => Saml::Elements::AttributeValue do
   end
 
   factory :attribute_statement, :class => Saml::Elements::AttributeStatement do